refactor: channel service에서 linkChannelToUserService로 코드 나눔#196
- channel service 에서 LinkChannelToUser repository 에 대한 소스의 양이
  많아져서 소스를 분리함.
- 포스트맨으로 모든 채널 api 1차적으로 테스트 완료
