+++ back/web-app/src/channels/channels.controller.ts
@@ -3,11 +3,13 @@ import { LinkChannelToUser } from './channel/entities/link-channel-to-user.entit
 import { MyChannels } from './channel/dto/my-channels.dto';
 import { ChannelService } from './channel/channel.service';
 import { Channel } from './channel/entities/channel.entity';
+import { LinkChannelToUserService } from './channel/link-channel-to-user.service';
 
 @Controller('channels')
 export class ChannelsController {
 	constructor(
 		private channelService: ChannelService,
+		private linkService: LinkChannelToUserService,
 	) {}
 
 	@Get()
@@ -19,7 +21,7 @@ export class ChannelsController {
   async getMyChannels(@Session() session: Record<string, any>,)
 	:Promise<MyChannels[]> {
     const userId: number = session.userId;
-    const links: LinkChannelToUser[] = await this.channelService.getLinksRelatedChannelByUserId(userId);
+    const links: LinkChannelToUser[] = await this.linkService.getLinksRelatedChannelByUserId(userId);
 
 		return await this.channelService.getMyChannels(links);
   }
